+++ CHANGELOG.md
@@ -5,7 +5,7 @@ documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [unreleased]
+## [0.0.41] - 2020-06-21
 ### Added
 - Add `-c <name>=<value>` option to pass a configuration parameter to the
   command.
@@ -396,4 +396,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 [0.0.38]: https://github.com/ktomk/pipelines/releases/tag/0.0.38
 [0.0.39]: https://github.com/ktomk/pipelines/releases/tag/0.0.39
 [0.0.40]: https://github.com/ktomk/pipelines/releases/tag/0.0.40
+[0.0.41]: https://github.com/ktomk/pipelines/releases/tag/0.0.41
 [unreleased]: https://github.com/ktomk/pipelines